Move onnx test to new file
